+++ include/caviar/objects/writer/force_field.h
@@ -35,9 +35,9 @@ class Force_field : public Writer {
   bool read (class caviar::interpreter::Parser *);
   void initialize();
   void write();
-  void write(int); // current time_step
+  void write(int64_t); // current time_step
   void write(double); // current time
-  void write(int, double); //time_step and time
+  void write(int64_t, double); //time_step and time
   void start_new_files(); //add_time_to_previous
   void start_new_files(std::string &); //add_time_to_previous
