+++ src/caviar/objects/writer.cpp
@@ -35,9 +35,9 @@ void Writer::verify_settings () {
 
 void Writer::initialize(){}
 void Writer::write(){}
-void Writer::write(int){} // current time_step
+void Writer::write(int64_t){} // current time_step
 void Writer::write(double){} // current time
-void Writer::write(int, double){} //time_step and time
+void Writer::write(int64_t, double){} //time_step and time
 void Writer::start_new_files(){} //add_time_to_previous
 void Writer::start_new_files(std::string &){} //add_time_to_previous
 void Writer::open_files() {}
